requirements.txt
streamlit
langchain
openai
duckduckgo-search
python-dotenv
twilio
transformers
torch
